Move capture_serverless to contrib/serverless
We're going to need context helpers and the like, makes sense
for it to live in its own area
